% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{choice_labels_to_names}
\alias{choice_labels_to_names}
\title{Convert all choice data from labels into names}
\usage{
choice_labels_to_names(object, language = NULL)
}
\description{
Convert all choice data from labels into names
}
