% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{choice_names_to_labels}
\alias{choice_names_to_labels}
\title{Convert all choice data from names into labels}
\usage{
choice_names_to_labels(object, language = NULL)
}
\description{
Convert all choice data from names into labels
}
