% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_converter.R
\name{class_converter}
\alias{class_converter}
\title{Function for converting all columns in a data frame to the proper type}
\usage{
class_converter(env)
}
\description{
Function for converting all columns in a data frame to the proper type
}
