% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_code_translation.R
\name{convert_calc}
\alias{convert_calc}
\title{Interpret calculations  from XLS Form}
\usage{
convert_calc(string)
}
\arguments{
\item{string}{String to be interpreted, should be of the standard format for calculate logic within XLS Forms.}
}
\description{
\code{convert_relevant} converts calculations logic from XLS Form into an R language equivalent.
}
\details{
Takes in a string of calculations from XLS Form and converts all portions of it into a parsed expression that
  can be passed to \code{dplyr::filter} for use in recalculating data after cleaning.
}
