% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_converter.R
\name{convert_columns}
\alias{convert_columns}
\title{Function for converting non-select multiple columns to the proper type}
\usage{
convert_columns(sheet, types, converter, env)
}
\description{
Function for converting non-select multiple columns to the proper type
}
