% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_code_translation.R
\name{convert_count_selected}
\alias{convert_count_selected}
\title{Interpret count-selected function from XLS Form}
\usage{
convert_count_selected(string)
}
\arguments{
\item{string}{String to be interpreted, should be of the format "count-selected($\{var\})"}
}
\value{
Returns a string
  value of the format `str_count(var, ' ')`, which can be parsed for use in filtering
  datasets. Works together with other internal functions to fully intrepet relevant
  logic strings from XLS Form coding in `convert_relevant()` and `convert_calculate()`.
}
\description{
`convert_count_selected` converts count-selected($\{var\}) into an equivalent R statement.
}
\examples{
convert_count_selected("count-selected(${livelihoods})")

}
