% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_code_translation.R
\name{convert_logical_blanks}
\alias{convert_logical_blanks}
\title{Interpret logicals involving blanks}
\usage{
convert_logical_blanks(string)
}
\arguments{
\item{string}{String to be interpreted, should be of the format "convert_logical_blanks($\{var\} = '')"}
}
\value{
Returns a string
value of the format `is.na(var)` or `!is.na(var)`. Works together with other internal functions to fully intrepet calculate
strings from XLS Form coding in `convert_calculate`.
}
\description{
`convert_logical_blanks` converts logical statements involving blanks into equivalent R statements using is.na.
}
\examples{
`convert_logical_blanks("${hh_needs} = ''")`

}
