% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevant_translation.R
\name{convert_relevant}
\alias{convert_relevant}
\title{Interpret relevant logic from XLS Form}
\usage{
convert_relevant(string)
}
\arguments{
\item{string}{String to be interpreted, should be of the standard format for relevant logic within XLS Forms.}
}
\description{
`convert_relevant` converts relevant relevant logic from XLS Form into R language.
}
\details{
Takes in a string of relevant logic from XLS Form and converts all portions of it into returns a parsed expression that
can be passed to `dplyr::filter` and used to filter data frames.
}
