% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_converter.R
\name{convert_select_multiple}
\alias{convert_select_multiple}
\title{Function for converting columns of select multiple individual options to logical vectors}
\usage{
convert_select_multiple(sheet, env)
}
\description{
Function for converting columns of select multiple individual options to logical vectors
}
