% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevant_translation.R
\name{convert_selected}
\alias{convert_selected}
\title{Interpret selected function from XLS Form}
\usage{
convert_selected(string)
}
\arguments{
\item{string}{String to be interpreted, should be of the format "selected(${var}, val)"}
}
\value{
String value of the format var %in% 'val', which can be
  parsed for use in filtering datasets. Works together with other internal functions
  to fully intrepet relevant logic strings from XLS Form coding in `convert_relevant`.
}
\description{
`convert_selected` converts selected(var, val) into an equivalent R statement.
}
