% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_code_translation.R
\name{convert_selected_at}
\alias{convert_selected_at}
\title{Interpret selected-at function from XLS Form}
\usage{
convert_selected_at(string)
}
\arguments{
\item{string}{String to be interpreted, should be of the format "count-selected(${var})"}
}
\value{
Returns a string
value of the format str_count(var, ' '), which can be parsed for use in filtering
datasets. Works together with other internal functions to fully intrepet strings from
XLS Form coding in `convert_relevant` and `convert_calculate`.
}
\description{
`convert_selected_at` converts selected-at(${var}, n) into an equivalent R statement.
}
\examples{
convert_selected_at("selected_at(${biggest_needs}, 1)")

}
