% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_code_translation.R
\name{convert_xls_code}
\alias{convert_xls_code}
\title{Interpret relevant logic from XLS Form}
\usage{
convert_xls_code(string)
}
\arguments{
\item{string}{String to be interpreted, should be of the standard format for relevant logic within XLS Forms.}
}
\description{
\code{convert_relevant} converts relevant relevant logic from XLS Form into an R language equivalent.
}
\details{
Takes in a string of relevant logic from XLS Form and converts all portions of it into a parsed expression that
  can be passed to \code{dplyr::filter} for use in filtering data frames.
}
\examples{
convert_relevant("selected(${employed_by_empire}, 'yes')")
convert_relevant("selected(${employed_by_empire}, 'yes') and count-selected(${livelihoods_coping}) > 3")

}
