% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{filter_expr_generator}
\alias{filter_expr_generator}
\title{Generate filter expression for cleaning functions}
\usage{
filter_expr_generator(index, uuid, relevant)
}
\description{
Generate filter expression for cleaning functions
}
