% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_identifiers.R
\name{identify_list_name}
\alias{identify_list_name}
\title{Funtion to identify list_name for select questions}
\usage{
identify_list_name(env)
}
\description{
Funtion to identify list_name for select questions
}
