% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobold_class.R
\name{kobold}
\alias{kobold}
\title{Create a kobold object from data frames}
\usage{
kobold(survey, choices = NULL, data = NULL, cleaning = NULL, ...)
}
\arguments{
\item{survey}{Data frame containing XLSForm survey.}

\item{choices}{Data frame containing XLSForm choices, optional.}

\item{data}{Data frame containing main XLSForm data export, optional.}

\item{cleaning}{Data frame containing cleaning sheet for use with kobold, optional.}

\item{...}{Data frame(s) for repeat group data; names must match repeat_group names from \code{survey}, optional.}
}
\description{
Constructor of a kobold list using XLSForm survey/choices and other possible data frames:
\itemize{
\item Main data sheet
\item Loop data sheet(s), which do not need to be explicitly loaded
\item Cleaning sheet
}
}
