% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_form.R
\name{read_xls_form}
\alias{read_xls_form}
\title{Load XLS Form file}
\usage{
read_xls_form(filepath, survey = "survey", choices = NULL,
  data = NULL, cleaning = NULL)
}
\arguments{
\item{filepath}{Path to the XLS Form}

\item{survey}{Name of the survey worksheet, Default: 'survey'}

\item{choices}{Name of the choices worksheet, optional, Default: 'choices';}

\item{data}{Name of the data worksheet, Default: 'data'}

\item{cleaning}{Name of the cleaning worksheet, Default: 'cleaning'}
}
\value{
Returns a \code{kobold} class object, a modified list containing the above data frames.
}
\description{
\code{read_xls_form} loads an XLS Form and its accompanying data into R.
}
\details{
This function loads XLS Form Excel files into R, while also allowing additional sheets for exported data and cleaning.
  It's loaded as an S3 class \code{kobold}, and can be worked with using functions within this package, or the data can
  be retrieved as a data frame or tibble using generics. The columns of the data are given a class according to the
  survey and choices sheet.
}
