% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_relevants.R
\name{same_relevants}
\alias{same_relevants}
\title{Function to update relevant logic referencing the same sheet}
\usage{
same_relevants(sheet, q_name, relevant, env)
}
\description{
Function to update relevant logic referencing the same sheet
}
