% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_relevants.R
\name{separate_relevants}
\alias{separate_relevants}
\title{Function to update relevant logic referencing a different sheet}
\usage{
separate_relevants(original_sheet, sheet, q_name, relevant, env)
}
\description{
Function to update relevant logic referencing a different sheet
}
