% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_expr_generator}
\alias{filter_expr_generator}
\title{Generate filter expression}
\usage{
filter_expr_generator(var_index = NA, var_uuid = NA, relevant = NA)
}
\description{
Generate filter expression
}
