% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobold_class.R
\name{new_kobold}
\alias{new_kobold}
\title{Create a kobold list from an XLS Form}
\usage{
new_kobold(data, cleaning, survey, choices)
}
\arguments{
\item{data}{Data frame containing exported data}

\item{cleaning}{Data frame containing cleaning sheet}

\item{survey}{Data frame containing XLS Form survey}

\item{choices}{Data frame containing XLS Form choices}
}
\description{
This is a constructor of a kobold list using 4 components of an XLS Form survey, data outputs, and the kobold cleaning sheet.
}
