% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevant_translation.R
\name{var_extract}
\alias{var_extract}
\title{Extract variable names from XLS Form}
\usage{
var_extract(string)
}
\arguments{
\item{string}{String to be interpreted, should be of the format "${var}"}
}
\value{
String value of the variable, which can be parsed for use in filtering datasets.
  Works together with other internal functions to fully interpret relevant logic string from XLS Form
  coding in `convert_relevant`.
}
\description{
`convert_selected` extracts the variable from the XLS Form
  format of ${variable}.
}
